add a separate variable for the runner agent private address
Allow separate control of the private addresses for the runners
and the Gitlab agent.

You might want to have separate auto assign ip address settings
for the Gitlab agent and runners.

As an example when you which to have the Gitlab agent with a
public ip address so it can register itself to the Gitlab
cloud. Were you want the runners to be in a private subnet
and have the external ip of the NAT gateway instead of their
own ip.

This is needed when you have a CICD account which will
run deployments in a separate account. The CICD account will
host the custom Gitlab runner.
In this separate account you want to whitelist the
CICD ip addresses so it can perform updates to the infrastructure.