allow the executor subnets to be specified independently
The runner and the executor do not need to be hosted in the same availability zone.

As long as proper routing is configured we could host the runner (manager) in "b",
while the executor is launched in "c".

By extending the runner configuration file with more [[runners]] entries we can
launch executors in multiple availability zones from a single runner. This is
especially convenient when using EC2 Spot instances for the executors as the
availability of specific machine types in availability zones varies.