refactor: make the agent termination lifecycle hook mandatory (#628)
Refactors the lambda function which handles killed agents.

rename it to terminate-agent-hook to make clear that it deals with the agents and not with the executors

make the function mandatory so it always runs as soon as an agent is terminated. This is needed in preparation for Remove resources created during runtime (housekeeping) #623 as we want to remove unused resources. So the function has to run
remove all variables which should be better managed by the module and not by the user, e.g. memory size and timeouts. They depend on the module only, so they should be managed by the module.

The variables which are no longer in use are still there, but have been marked as deprecated. So no migration is necessary.