fix: install gitlab-runner after docker+machine driver (#704)
## Description

As pointed out in #703 the GitLab Runner is installed and started before the docker+machine driver. But as the docker+machine driver needs some time for a complete installation (certificates, ...) it might happen that the GitLab Runner already tries to process jobs from GitLab. These jobs will fail.

This PR changes the installation order of the GitLab Runner project and the docker+machine driver. The driver is installed first, the GitLab Runner second. This avoids running a non-working GitLab Runner setup.

⚠️ The restart of the GitLab Runner after executing user's post install script has been removed as a restart could possibly loose some jobs being processed by the GitLab Runner. If this is needed by your script, you have to add it there and be aware, that this might cause problems with the first jobs being processed.

Closes #703