ci: lint all files with Megalinter (#580)
## Description

This PR makes sure that all files are syntactically valid and adhere to
the specific format. This makes it easier to read the files in the
repository and that the code follows the industry standard.

[Megalinter](https://github.com/oxsecurity/megalinter) is used to have
all relevant linters at hand. But some of them were disabled:
- Trivy: Nice linter to report CVEs and other cool stuff. But it reports
problems with the Terraform code which can't be disabled by
configuration
- TfLint: Has some problems reading referenced policy files. I created a
separate workflow for TfLint as it creates valuable output.
- Markdown table formatter: The tables created by TfDoc are not
formatted according to the standard. And there is no option to
deactivate the check for a section of the file.
- Misspell: CSpell does a great job. No need for a second linter

## Migrations required

No

## Verification

Verified that the job runs and throws an error if a markdown file is not
valid.