ci: run KICS in parallel to Megalinter to gain performance (#755)
## Description

This PR runs KICS in parallel to the Megalinter. KICS is a long running
linter (up to 4 minutes). We can save roughly 3 minutes if we run it in
parallel to Megalinter. So we don't have to wait for Docker to download
the Megalinter image.

KICS finishes now before the Megalinter.

## Migrations required

No

## Verification

Check that the workflow is green.