fix: disable outputting config.toml by default (#768)
## Description

In #754 I unfortunately made a major oversight: the `local_file` doesn't
persist between machines (duh 🤦🏾 ). So for example in a CI/CD pipeline
the file is not present so it wants to render the whole file as new
every time. This defeats the main purpose that I had intended for this
feature... Also if the pipeline applies the plan, there will be a
perpetual diff, even if you save the file in repository.

I still think this is useful locally for debugging purposes, but
therefore I propose we disable this mechanism by default. In addition,
when these variables are enabled I changed the location to the Root
module for easier access if you want to debug the rendered files.

## Migrations required

YES/NO - Feature is quite new, if someone wants to use keep this as
introduced they now need to explicitly enable it.

## Verification

```hcl
output_runner_config_to_file = true
output_user_data_to_file     = true
```

Results:


![image](https://user-images.githubusercontent.com/17970041/227529895-864e45e0-70aa-4b23-a725-30c4cea8b955.png)

---------

Co-authored-by: kayma <kayma@hlag.com>
Co-authored-by: Matthias Kay <github-public@matthiaskay.de>