revert: removes the `nonsensitive` from `runner_user_data` output (#832)
## Description

Removes the `nonsensitive` function as the value sometime is not
sensitive and the function fails with

```
│ Error: Invalid function argument
│
│ on .terraform/modules/gitlab_runner_sare_pim_infrastructure/outputs.tf line 58, in output "runner_user_data":
│ 58: value = nonsensitive(local.template_user_data)
│ ├────────────────
│ │ while calling nonsensitive(value)
│ │ local.template_user_data is
```

The output is marked as deprecated and shouldn't be used any longer. Use
`var.debug.output_runner_user_data_to_file = true` to write the user
data as file to the local disk.

Closes #828 

## Migrations required

No

## Verification

No verification done.