feat: add `MaxGrowthRate` to limit the number of instances added in parallel (#962)
## Description

This feature allows to support `MaxGrowthRate` which limits the number
of instances added in parallel by the Runner. The default of `0`
(unlimited) is used if no value is given.

The parameter is added to `runner_worker_docker_machine_instance` as
this is the only Runner type for which it makes sense. All other types
are not able to scale and thus it is not a `runner_worker` parameter.

See
[advanced-configuration](https://docs.gitlab.com/runner/configuration/advanced-configuration.html#the-runnersmachine-section)

Closes #851 

## Migrations required

NO

## Verification

- do not set the parameter --> `MaxGrowthRate` in config file is 0.
- set the parameter to any value --> `MaxGrowthRate` in config file is
"any value"

---------

Co-authored-by: Matthias Kay <github@matthiaskay.de>
Co-authored-by: Matthias Kay <matthias.kay@hlag.com>