chore: fix minimum Terraform version for TfEnv (#1001)
## Description

Make sure that we have a `.terraform-version` file for all examples to
be able to install Terraform with `tfenv install`. As the version number
in this file indicates the minimum version to use, we don't want any
updates from Renovate.

All examples used the wrong minimum Terraform version as we raised it in
the main module to 1.3.