fix: delete the 'runner_config_toml_rendereded' output variable (#1019)
## Description

Due to calls to `nonsensitive` in case the data is sensitive or not
calling the function in the other case, the code gets very complex. The
output variable `runner_config_toml_rendereded` was introduced for
easier debugging, but we still have the option to write the whole Runner
configuration to a local file (`debug.write_runner_config_to_file`).

This "fixers" that issue by just deleting the output variable.

Closes #1018

## Migrations required

Usually not. For debugging purposes, set the
`debug.write_runner_config_to_file` to `true`. The configuration is
written to a file on your local disk (directory `debug/`)

## Verification

Only linted as this is only deleting an output variable.