test: add unit test framework for Terraform (#1038)
## Description

This PR adds the possibility to add automated tests for the module. We
start simple with a `terraform test` command. The tests live at
`tests/`.

Tests are executed only if the base branch was created in the cattle-ops
repository. Terraform needs cloud access which is restricted to the
cattle-ops organization at the moment. The AWS account is sponsored by
Hapag-Lloyd.

## Migrations required

No

## Verification

No verification done as this PR adds tests only.