fix: remove the deprecated runner_user_data output variable (#1032)
## Description

Depending on the value of the output a call to `nonsensitive` is needed.
Instead of introducing a conditional call, we remove the output
variable. The value is still written to a local file, see
`debug.write_runner_user_data_to_file`.

Closes #1031.

## Migrations required

YES - any use of the `runner-user-data` output must be changed to use
`var.debug.write_runner_user_data_to_file`. As this is used for debug
only, it shouldn't be a problem for your live environment.

## Verification

None. Just removing the output variable.

Co-authored-by: Ryan Causey <ryan.causey@munipal.io>
Co-authored-by: Matthias Kay <matthias.kay@hlag.com>