fix: rename `pull_policies` to `pull_policy` in `config.toml` (#1084)
## Description

The `runner_worker_docker_options` mentioned a `pull_policies` attribute
which was directly used in `config.toml` which is wrong. The correct
name is `pull_policy`. `pull_policy` is ignored by the Runner and has no
effect. Instead a pull policy of `always` is used automatically.

This PR renames the `pull_policies` to `pull_policy` before inserting it
into the configuration file. If `pull_policies` is used at the moment it
has no effect.

Closes #1065 

## Migrations required

In case `var.runner_worker_docker_options.pull_policy` is used, rename
`pull_policy` to `pull_policies` and make sure that it is a
`list(string)`, e.g. `pull_policies = ["always"]`

## Verification

Still to be done

- [X] `pull_policies` not given --> `pull_policy = ["always"] is used.
- [X] `pull_policies` given --> Specified value used