fix: do not deregister GitLab tokens which are still in use (#1102)
## Description

So far we haven't tracked the usage of the GitLab Runner token. In case
the Runner is stopped and a new one is spawned, race conditions might
occur as explained in #1062. In consequence the new Runner used a token
deleted by the old Runner. This leads into downtimes as the token is no
longer valid and can't be used.

This PR converts the token into a JSON format and ensures that tokens
which are still in use, are not deregistered.

Fixes #1062 

## Migrations needed

In case you want to rollback to a previous version you have to convert
the SSM parameter containing the token and usage counter in JSON format
back to a plain token string.

## Verification

- [x] script changes were tested locally
- [x] Test Runner was started to ensure that the token conversion works

---------

Co-authored-by: long-wan-ep <50891790+long-wan-ep@users.noreply.github.com>