fix: use existing Runner package for Amazon Linux 2023 as default (#1155)
## Description

Since #1138 was merged, the `runner_gitlab.runner_version` was wrong.
The mentioned package `15.8.2` does not exist for Amazon Linux 2023.
This PR updates the default version to `16.0.3` the oldest available
version.

You should never rely on the default value as your Runner has to match
the GitLab version. See
https://docs.gitlab.com/runner/#gitlab-runner-versions

Closes #1147 

## Migrations required

Usually not. In case you are still on GitLab < 16 you have to use the
old AMI matching `amzn2-ami-hvm-2.*-x86_64-ebs` (used before #1138 was
merged). Then you are able to still use an older version of the Runner.

## Verification

- verified that `16.0.3` is available. Installation is done via
```
  curl --fail --retry 6 -L https://packages.gitlab.com/install/repositories/runner/gitlab-runner/script.rpm.sh | bash
  yum install gitlab-runner-${gitlab_runner_version} -y
```