feat: add `pre_get_sources_script` for GitLab Runner 17+ (#1158)
## Description

Adds the `pre_get_sources_script` (GitLab Runner 17+) parameter which is
filled with the value of `pre_clone_script` (GitLab Runner < 17).

The parameter was renamed. Deprecation warnings are shown in the logs of
GitLab Runner 16. GitLab Runner 17 simply ignores the parameter and
therefor the whole script. This might result in failing pipelines.

## Verification

- [x] use Runner 16 and verify that the script is taken into account
- [x] use Runner 17 and verify that the script is taken into account
- [x] no errors and warnings are shown in the log