feat: add docker autoscaler executor (#1118)
## Description
 
Provides a new executor using the new GitLab autoscaler executor. I've
been using the fleeting plugin for AWS only.

Prerequisite: Docker must already be installed on the AMI used by worker
machines (the Docker autoscaler does not install it, unlike the Docker
machine). Additionally, the user used to connect to the workers must
also be added to the Docker group.

Related to issue
https://github.com/cattle-ops/terraform-aws-gitlab-runner/issues/624

## Verification

Built an AMI with Docker based on Amazon Linux 2023. Set up the new
executor according to the example. Works!

---------

Co-authored-by: Matthias Kay <matthias.kay@hlag.com>