feat: make the graceful terminate Lambda timeout configurable (#1164)
## Description

Add `var.runner_terminate_ec2_timeout_duration` with a default of 90
seconds. Makes it possible to configure the timeout in case hundreds of
workers are controlled by a single GitLab Runner Agent. Otherwise the
Lambda function might fail and does not remove any worker in case the
Agent is stopped.

Closes #1149