fix: use correct names for autoscaler policies configuration (#1167)
## Description

PascalCase was used to create the `[[runners.autoscaler.policy]]` which
is wrong according to the documentation. This PR fixes the names in the
configuration file. `IdleCount` --> `idle_count`

The worker instances were used once and then terminated. This resulted
in a high load of some AWS autoscaling APIs and we got rate limiting. In
the end most pipelines failed.

## Verification

Deployed the changes into my environment. The worker machines are reused
now.