Allow changes to "runner_worker.max_jobs" for Docker Autoscaler
Remove ignore_changes from the ASG for Docker Autoscaler.  The reasoning
described in the comment was poor; min_size and desired_capacity are fixed
at zero, so won't change, and max_size should change when runner_worker.max_jobs
is changed, and will not cause immediate scale-up.
If it causes scale-down, this the logical desired behaviour.