fix: allow changes to "runner_worker.max_jobs" for Docker Autoscaler (#1221)
## Description

Remove `min_size` and `max_size` from `ignore_changes` from the ASG for
Docker Autoscaler. The reasoning described in the comment was poor;
min_size is fixed at zero, so won't change, and max_size should change
when runner_worker.max_jobs is changed, and will not cause immediate
scale-up. If it causes scale-down, this the logical desired behaviour.

---------

Co-authored-by: Matthias Kay <matthias.kay@hlag.com>