feat: allow instrumentation of Termination lambda (#1255)
## Description

This PR adds the possibility to instrument the internal Lambda function,
e.g. with APM tools. Use the following variables:

- `runner_terminate_ec2_lambda_handler` to replace the `handler` with
your function
- `runner_terminate_ec2_environment_variables` to add environment
variables. The special value `{HANDLER}` is automatically replaced by
the internal handler name to be able to call the "real" handler
- `runner_terminate_ec2_lambda_handler_layer_arns` to add additional
layers to the Lambda function
- `runner_terminate_ec2_lambda_egress_rules` to allow traffic to
external systems. IPv4/6 port 443 is the default