chore: make the termination handler more robust (#1322)
## Description

In case the agent is terminated, we search for child instances which are
`'running', 'pending', 'stopping', 'stopped'`. All ol them are
terminated. Furthermore we retry the termination call 3 times in case of
exceptions.

This fixes the problem that sometimes a small amount ol EC2 instances
remain in stopped state.

Fixes parts of #1298