feat: add `connection_max_age` config (#1340)
## Description

Currently seeing some occasional connectivity issues running behind a
NAT gateway - where connections are limited to 5 minutes. The
gitlab-runner config by defaults limits connections to 15 minutes. Was
looking to set the connection_max_age setting with the module, but could
not find an option to do so.

Add connection_max_age as a configurable option to the runner config.

Co-authored-by: Matthias Kay <matthias.kay@hlag.com>